+++ NEWS.md
@@ -5,6 +5,8 @@
 ## alced.api version 1.1.3
 * Added functionality to do immediate geographic/temporal aggregation
 * Resized README example data retrieval
+* Updated region names and codes based on new ACLED API guide
+* Bug fix in region call
 * 1 merge request by Rob Williams https://github.com/jayrobwilliams 
 + !3 Bug fix: %20 instead of % for white spaces in country names
  -->